Let malloc infer its attributes
Malloc should not force '@safe' or '@nogc' on constructors which are not.
For example, TaskPool's ctor is not '@nogc' but was assumed as such
thanks to the delegate cast happening in malloc.
Likewise, the ctor or the arguments might not be '@safe',
or any other attributes, but they were be mistakenly marked as such.