Revert "Work around possible hang of CFRunLoop with pending kqueue events."
This reverts commit decbe7f6bdeb292183ac447c853d5825eacca8ff.

Using level triggered events resulted in 100% CPU usage, because sockets receive a constant write-ready event. They register themselves with the event loop only once in the beginning for efficiency reasons and thus may receive these events without any active waiters.

In case of kqueue it would be possible to register dynamically only when waiting, without a big performance impact, but in case of the epoll API this would result in am excessive growth of kernel calls.